requirements.txt
imageio==2.27.0
numpy==1.24.2
Pillow==9.5.0
pygifsicle==1.0.7
support-developer==1.0.5
